refactor: changed project structure to nest architecture
