+++ README.md
@@ -26,7 +26,7 @@ npm start
 After starting the app on port (4000 as default) you can open
 in your browser OpenAPI documentation by typing http://localhost:4000/doc/.
 For more information about OpenAPI/Swagger please visit https://swagger.io/.
-
+To change the port, open .env file and change the port value.
 ## Testing
 
 After application running open new terminal and enter:
