+++ src/api/album/interface/album.interface.ts
@@ -1,0 +1,8 @@
+import { UUID } from 'crypto';
+
+export interface IAlbum {
+  id: UUID;
+  name: string;
+  year: number;
+  artistId: string | null; // refers to Artist
+}
