+++ src/api/artist/interface/artist.interface.ts
@@ -1,0 +1,7 @@
+import { UUID } from 'crypto';
+
+export interface IArtist {
+  id: UUID;
+  name: string;
+  grammy: boolean;
+}
