+++ src/api/favorites/favorites.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { FavController } from './favorites.controller';
+import { FavService } from './favorites.service';
+
+@Module({
+  controllers: [FavController],
+  providers: [FavService],
+})
+export class FavModule {}
