+++ src/api/track/interface/track.interface.ts
@@ -1,0 +1,9 @@
+import { UUID } from 'crypto';
+
+export interface ITrack {
+  id: UUID;
+  name: string;
+  artistId: string | null; // refers to Artist
+  albumId: string | null; // refers to Album
+  duration: number; // integer number
+}
